+++ data/duke.txt
@@ -1,2 +1,2 @@
-1.T|NotDone a 
-2.E|NotDone a (by: sdklfj)
+1.T|Done a 
+2.T|Done a
